requirements.txt
# Core application and web framework
streamlit
python-dotenv
selenium
# Google Cloud & AI Services
google-genai
google-adk
serpapi

# Data Scraping and Parsing
beautifulsoup4
yt-dlp
# selenium  <-- See WARNING below

# Database Driver (can be removed if using the line above)
psycopg2-binary
lxml
google-search-results
serpapi